# Implement one shot SHA and reduce allocs