--- a/openssl/sha_test.go
+++ b/openssl/sha_test.go
@@ -66,6 +66,47 @@ func TestSha(t *testing.T) {
 	}
 }
 
+func TestSHA1(t *testing.T) {
+	msg := []byte("testing")
+	var tests = []struct {
+		name string
+		fn   func() hash.Hash
+		fn2  func([]byte) []byte
+	}{
+		{"sha1", NewSHA1, func(p []byte) []byte {
+			b := SHA1(p)
+			return b[:]
+		}},
+		{"sha224", NewSHA224, func(p []byte) []byte {
+			b := SHA224(p)
+			return b[:]
+		}},
+		{"sha256", NewSHA256, func(p []byte) []byte {
+			b := SHA256(p)
+			return b[:]
+		}},
+		{"sha384", NewSHA384, func(p []byte) []byte {
+			b := SHA384(p)
+			return b[:]
+		}},
+		{"sha512", NewSHA512, func(p []byte) []byte {
+			b := SHA512(p)
+			return b[:]
+		}},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := tt.fn2(msg)
+			h := tt.fn()
+			h.Write(msg)
+			want := h.Sum(nil)
+			if !bytes.Equal(got[:], want) {
+				t.Errorf("got:%x want:%x", got, want)
+			}
+		})
+	}
+}
+
 func BenchmarkHash8Bytes(b *testing.B) {
 	b.StopTimer()
 	h := NewSHA256()
@@ -79,5 +120,17 @@ func BenchmarkHash8Bytes(b *testing.B) {
 		h.Reset()
 		h.Write(buf[:size])
 		h.Sum(sum[:0])
+	}
+}
+
+func BenchmarkSHA256(b *testing.B) {
+	b.StopTimer()
+	buf := make([]byte, 8192)
+	size := 1024
+	b.StartTimer()
+	b.SetBytes(int64(size))
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		SHA256(buf[:size])
 	}
 }