This PR contains two performance improvements which landed into `dev.boringbranch` during this development cycle, both committed in https://go-review.googlesource.com/c/go/+/395876/11.

- a99ec67c99c0e0d2bc69e495dbb0794969ec3314: Implement one shot SHA calls. This is helpful for callers which don't need to keep the hash object returned by `NewSha1` and friends, as these functions return a pointer which normally escapes in the caller side.
- 9fafecd26c83a05018bd5a64541c488c2cb247db and 31857f33f60aef96df3b40722c163ebdf13b6fc5: Makes SHA calls allocation-free by hiding openssl pointers from the escape analysis.

The benchmarks demonstrates that the new code does not allocate:

```cmd
name          old time/op    new time/op    delta
Hash8Bytes-4    4.57µs ± 4%    4.46µs ± 4%    -2.55%  (p=0.009 n=10+10)

name          old speed      new speed      delta
Hash8Bytes-4   224MB/s ± 4%   230MB/s ± 4%    +2.62%  (p=0.009 n=10+10)

name          old bytes/op   new bytes/op   delta
Hash8Bytes-4     24.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)

name          old allocs/op  new allocs/op  delta
Hash8Bytes-4      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
```

Removing the SHA allocations is important for when openssl is integrated into go1.19, else we would have to disable a new test that checks boring SHAs don't allocate.