@qmuntal 2022-05-12
I've found something good: `noescapeCtx` is not necessary because we are passing OpenSSL C pointers as `uintptr` instead of `unsafe.Pointer`, and the former are already safely hidden from the escape analysis.